split up colorHelpers and colorSequenceHelpers
