+++ src/Form.tsx
@@ -3,7 +3,8 @@ import CheckboxInput from './inputs/Checkbox'
 import TogglableColorPicker from './inputs/TogglableColorPicker'
 import IntegerInput from './inputs/Integer'
 import { Color, SwatchConfig } from './types'
-import { getRandomNotWhiteColor, totalColorSequenceLength } from './colorHelpers'
+import { getRandomNotWhiteColor } from './colorHelpers'
+import { totalColorSequenceLength } from './colorSequenceHelpers'
 import { mod } from './numberHelpers'
 
 type FormValue = keyof(SwatchConfig)
